add: TSAnalysis
