+++ ElasticNetVAR/coordinate_descent.jl
@@ -1,4 +1,37 @@
 """
+    ijth_coordinate_update!(i::Int64, j::Int64, T_minus_p::Int64, estim_settings::EstimSettings, Y_i::SubArray{Float64}, Y_lagged::FloatArray, Ψ::FloatArray)
+
+Update the (i,j)-th element of Ψ the coordinate descent algorithm (Friedman et al., 2010) as in Pellegrino (2019).
+
+# References
+Friedman et al. (2010) and Pellegrino (2019)
+"""
+function ijth_coordinate_update!(i::Int64, j::Int64, T_minus_p::Int64, estim_settings::EstimSettings, Y_i::SubArray{Float64}, Y_lagged::FloatArray, Ψ::FloatArray)
+
+    # Views
+    Y_lagged_j = @view Y_lagged[j,:];
+
+    # Partial residuals
+    Ψ_i_ex_j = Ψ[i,:]; # Julia creates a copy of Ψ[i,:]
+    Ψ_i_ex_j[j] = 0.0;
+    V_ij = Y_i - Y_lagged'*Ψ_i_ex_j;
+
+    # Scalar product between Y_lagged_j and V_ij divided by (T-p)
+    scalar_Y_lagged_j = sum(V_ij .* Y_lagged_j)/T_minus_p;
+
+    # Norm squared of Y_lagged_j divided by (T-p)
+    normsq_Y_lagged_j = sum(Y_lagged_j .^ 2)/T_minus_p;
+
+    # Soft-thresholding operator: update for the coefficients
+    hyper_prod = estim_settings.λ*estim_settings.β^fld(j-1, estim_settings.n);
+    Ψ[i, j] = soft_thresholding(scalar_Y_lagged_j, estim_settings.α*hyper_prod)/(normsq_Y_lagged_j + (1-estim_settings.α)*hyper_prod);
+
+    # Return update for objfun_new
+    update_objfun_new = hyper_prod*(0.5*(1-estim_settings.α)*Ψ[i,j]^2 + estim_settings.α*abs(Ψ[i,j]));
+    return update_objfun_new;
+end
+
+"""
     coordinate_descent(estim_settings::EstimSettings)
 
 Estimate an elastic-net VAR(p) with the coordinate descent algorithm (Friedman et al., 2010) as in Pellegrino (2019).
@@ -73,37 +106,4 @@ function coordinate_descent(estim_settings::EstimSettings)
 
     # Return output
     return Ψ, Σ;
-end
-
-"""
-    ijth_coordinate_update!(i::Int64, j::Int64, T_minus_p::Int64, estim_settings::EstimSettings, Y_i::SubArray{Float64}, Y_lagged::FloatArray, Ψ::FloatArray)
-
-Update the (i,j)-th element of Ψ the coordinate descent algorithm (Friedman et al., 2010) as in Pellegrino (2019).
-
-# References
-Friedman et al. (2010) and Pellegrino (2019)
-"""
-function ijth_coordinate_update!(i::Int64, j::Int64, T_minus_p::Int64, estim_settings::EstimSettings, Y_i::SubArray{Float64}, Y_lagged::FloatArray, Ψ::FloatArray)
-
-    # Views
-    Y_lagged_j = @view Y_lagged[j,:];
-
-    # Partial residuals
-    Ψ_i_ex_j = Ψ[i,:]; # Julia creates a copy of Ψ[i,:]
-    Ψ_i_ex_j[j] = 0.0;
-    V_ij = Y_i - Y_lagged'*Ψ_i_ex_j;
-
-    # Scalar product between Y_lagged_j and V_ij divided by (T-p)
-    scalar_Y_lagged_j = sum(V_ij .* Y_lagged_j)/T_minus_p;
-
-    # Norm squared of Y_lagged_j divided by (T-p)
-    normsq_Y_lagged_j = sum(Y_lagged_j .^ 2)/T_minus_p;
-
-    # Soft-thresholding operator: update for the coefficients
-    hyper_prod = estim_settings.λ*estim_settings.β^fld(j-1, estim_settings.n);
-    Ψ[i, j] = soft_thresholding(scalar_Y_lagged_j, estim_settings.α*hyper_prod)/(normsq_Y_lagged_j + (1-estim_settings.α)*hyper_prod);
-
-    # Return update for objfun_new
-    update_objfun_new = hyper_prod*(0.5*(1-estim_settings.α)*Ψ[i,j]^2 + estim_settings.α*abs(Ψ[i,j]));
-    return update_objfun_new;
 end
